Move is_dataframe_like and related functions to daks/utils.py
