+++ dask/utils.py
@@ -931,6 +931,31 @@ def is_arraylike(x):
     )
 
 
+def is_dataframe_like(df):
+    """ Looks like a Pandas DataFrame """
+    typ = type(df)
+    return (all(hasattr(typ, name)
+                for name in ('groupby', 'head', 'merge', 'mean')) and
+            all(hasattr(df, name) for name in ('dtypes',)) and not
+            any(hasattr(typ, name)
+                for name in ('value_counts', 'dtype')))
+
+
+def is_series_like(s):
+    """ Looks like a Pandas Series """
+    typ = type(s)
+    return (all(hasattr(typ, name) for name in ('groupby', 'head', 'mean')) and
+            all(hasattr(s, name) for name in ('dtype', 'name')) and
+            'index' not in typ.__name__.lower())
+
+
+def is_index_like(s):
+    """ Looks like a Pandas Index """
+    typ = type(s)
+    return (all(hasattr(s, name) for name in ('name', 'dtype'))
+            and 'index' in typ.__name__.lower())
+
+
 def natural_sort_key(s):
     """
     Sorting `key` function for performing a natural sort on a collection of
